Small update of stan_inference.jl to include Stan's exposed Adams-Moulton solver
Stan has had an Adams-Moulton method exposed for the past couple of years now as `integrate_ode_adams`, but I only recently found out because it was only explicitly mentioned in the latest update of their documentation. Refer to https://github.com/stan-dev/stan/issues/2542. Now recommended as the default non-stiff solver by people in the Stan community over rk45. Just a small suggested update to accommodate that as a result. Thanks to the DifferentialEquations.jl team for their hard work once more!